+++ index.html
@@ -30,7 +30,7 @@
       </div>
     </div>
 
-    <script src="dist/index.js"></script>
+    <script src="dist/cocoen.js"></script>
     <script>
       Cocoen.parse(document.body);
     </script>
