+++ src/config.ts
@@ -1,0 +1,2 @@
+export const name = 'cocoen';
+export const componentName = `${name}-component`;
