+++ src/create.ts
@@ -1,0 +1,20 @@
+import { componentName } from './config';
+import { Cocoen } from './components/cocoen';
+
+export const create = (element: Element): Cocoen => {
+  const component = document.createElement(componentName) as Cocoen;
+  const before = element.querySelectorAll('img')[0];
+  const after = element.querySelectorAll('img')[1];
+
+  if (!before || !after) {
+    throw new Error('Cocoen needs two images');
+  }
+
+  before.setAttribute('slot', 'before');
+  after.setAttribute('slot', 'after');
+  component.append(before.cloneNode(true));
+  component.append(after.cloneNode(true));
+  element.replaceWith(component);
+
+  return component;
+};
